backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
playwright==1.44.0
anthropic==0.28.0
supabase==2.5.0
python-dotenv==1.0.1
pydantic==2.7.0
python-multipart==0.0.9